backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
openai-whisper==20231117
torch==2.0.1
torchaudio==2.0.2
transformers==4.35.0
deepl==1.12.0
edge-tts==6.1.8
pydub==0.25.1
eventlet==0.33.3
numpy==1.24.3
librosa==0.10.1
soundfile==0.12.1
python-socketio==5.9.0
python-engineio==4.7.1
accelerate==0.24.1
sentencepiece==0.1.99
protobuf==4.25.0
requests==2.31.0
urllib3==2.0.7
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
click==8.1.7
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.3
werkzeug==2.3.7
bidict==0.22.1
dnspython==2.4.2
six==1.16.0
greenlet==3.0.1